refactor: updated pages and routes
