+++ src/routes/Routes.jsx
@@ -1,31 +1,15 @@
 import React from "react";
 import { Switch, Route } from "react-router-dom";
-import { arrayOf, func } from "prop-types";
-import { genreType, movieType, modalValues } from "types";
-import { Home, Movie } from "routes";
 import { PATHS } from "consts";
+import Home from "routes/Home";
+import Movie from "routes/Movie";
 
-const Routes = (props) => {
-  return (
-    <Switch>
-      <Route exact path={PATHS.HOME}>
-        <Home {...props} />
-      </Route>
-      <Route path={PATHS.MOVIE}>
-        <Movie {...props} />
-      </Route>
-      {/* <Route component={NotFound} /> */}
-    </Switch>
-  );
-};
+const Routes = () => (
+  <Switch>
+    <Route exact path={PATHS.HOME} component={Home} />
+    <Route path={PATHS.MOVIE} component={Movie} />
+    {/* <Route component={NotFound} /> */}
+  </Switch>
+);
 
-Routes.propTypes = {
-  genres: arrayOf(genreType),
-  movies: arrayOf(movieType),
-  modalValues: modalValues,
-  defaultModalValues: modalValues,
-  onModalValuesChange: func,
-  onMovieUpdate: func
-};
-
-export { Routes };
+export default Routes;
