+++ src/routes/index.js
@@ -1,3 +1,1 @@
-export { Routes } from "./Routes";
-export { Home } from "./Home";
-export { Movie } from "./Movie";
+export { default } from "./Routes";
